refactor: unificacion logica compresion imagen
- Se actualizó `src/utils/helpers.js` para usar `URL.createObjectURL` y `canvas` para la compresión de imágenes, soportando opciones de maxDimension, calidad y tipo.
- Se actualizó `src/modules/municipal.js` para usar la función centralizada `comprimirImagen`, preservando el comportamiento original (salida WebP, dimensión máxima de 1024px).
- Se corrigió un error en `municipal.js` donde las imágenes WebP se subían con el Content-Type original del archivo.
- Se eliminó la importación no utilizada de `comprimirImagen` en `src/modules/reports.js`.
